+++ Models/FundsModel.cs
@@ -8,10 +8,10 @@ namespace BrookingsIndoorTrainingSystem.Models
     public class FundsModel
     {
         public int id { get; set; }
-        public int Concessions { get; set; }
-        public int Equipment { get; set; }
-        public int Space { get; set; }
-        public int BITS { get; set; }
+        public double Concessions { get; set; }
+        public double Equipment { get; set; }
+        public double Space { get; set; }
+        public double BITS { get; set; }
 
     }
 }
